+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.assets.PoacController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.assets.ItemTypeIndicatorController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.assets.ItemLabelController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.assets.ItemController。
 
 #### Bug修复